change configuration

+++ secrets/dev-enableops.sh
@@ -1,9 +1,22 @@
 #!/usr/bin/env sh
 
-set -e 
+set -e
+
+if [ "$GCRJSONAUTH" = "" ]; then
+	echo "\nGCRJSONAUTH env must be provided"
+	exit 1
+fi
 
 kubectl create namespace enableops-api || true
+
 kubectl -n enableops-api delete secret python-env || true
+kubectl -n enableops-api delete secret eu-gcr-io || true
+
+kubectl -n enableops-api create secret docker-registry eu-gcr-io \
+  --docker-server=eu.gcr.io \
+  --docker-username=_json_key \
+  --docker-password="$GCRJSONAUTH" \
+  --docker-email=andrei@chenchik.me
 
 kubectl -n enableops-api create secret generic python-env \
 	--from-literal=API_SENTRY_DSN=$API_SENTRY_DSN \
@@ -15,4 +28,3 @@ kubectl -n enableops-api create secret generic python-env \
 	--from-literal=API_DB__HEROKU_API_KEY=$API_DB__HEROKU_API_KEY \
 	--from-literal=API_ENV_STATE=$API_ENV_STATE \
 	--from-literal=API_SECURITY__TERRAFORM_SECRET=$API_SECURITY__TERRAFORM_SECRET
-
